separate address info component
